refactor: colocate remaining app input tests
